Remove export methods from TableItem and VectorItem classes; refactor export logic to BaseItem class for better code organization.
